requirements.txt
altgraph==0.17.5
asgiref==3.11.0
bottle==0.13.4
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
macholib==1.16.4
packaging==25.0
proxy_tools==0.1.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1
pywebview==6.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
